refactor to decouple
