+++ ui/src/app/page.tsx
@@ -6,47 +6,45 @@ import { clsx } from 'clsx'
 import Board from '@/components/board/Board'
 import SearchBox from '@/components/search-box/SearchBox'
 import { Button, IconButton } from '@/components/button/Button'
-import PopUp from '@/components/pop-up/PopUp'
-import { Dropdown, DropdownItem } from '@/components/dropdown/Dropdown'
 import Tooltip from '@/components/tooltip/Tooltip'
 import { BookOpenIcon } from '@heroicons/react/24/outline'
 import { InformationCircleIcon } from '@heroicons/react/20/solid'
 import { Cog6ToothIcon } from '@heroicons/react/20/solid'
-import { usePlayers, UsePlayersReturn } from '@/hooks/usePlayers'
+import { usePlayers } from '@/hooks/usePlayers'
 import { useGame } from '@/hooks/useGame'
+import { PopupType } from '@/types/Popup'
+import { ReferencesPopup } from '@/components/pop-up/ReferencesPopup'
+import { HelpPopup } from '@/components/pop-up/HelpPopup'
+import { SettingsPopup } from '@/components/pop-up/SettingsPopup'
+import { GameOverPopup } from '@/components/pop-up/GameOverPopup'
 
 function App() {
     const [selectedEra, setSelectedEra] = useState<Era>(Era.ALL)
-    const { players }: UsePlayersReturn = usePlayers(selectedEra);
+    const { players } = usePlayers(selectedEra);
     const game = useGame({ players });
-
-    // pop-ups
-    const [isGameOverPopupActive, setGameOverPopupActive] = useState(false)
-    const [isHelpPopupActive, setHelpPopupActive] = useState(false)
-    const [isSettingsPopupActive, setSettingsPopupActive] = useState(false)
-    const [isReferencesPopupActive, setReferencesPopupActive] = useState(false)
+    const [activePopup, setActivePopup] = useState<PopupType>(PopupType.NONE);
 
     useEffect(() => {
         if (game.currAttempt === 7 && !game.gameOver) {
-            setGameOverPopupActive(true)
+            setActivePopup(PopupType.GAME_OVER)
         }
     }, [game.currAttempt, game.gameOver])
 
     return (
         <div className="font-mono text-center h-dvh w-dvw bg-white dark:bg-zinc-900 text-black dark:text-white/80 [&_*]:border-black/20 [&_*]:dark:border-white/20">
             <header className="absolute left-12 top-6 grid grid-cols-3 gap-2">
-                <Tooltip content="References">
-                    <IconButton fn={() => setReferencesPopupActive(true)}>
+                <Tooltip text="References">
+                    <IconButton fn={() => setActivePopup(PopupType.REFERENCES)}>
                         <BookOpenIcon />
                     </IconButton>
                 </Tooltip>
-                <Tooltip content="Help">
-                    <IconButton fn={() => setHelpPopupActive(true)}>
+                <Tooltip text="Help">
+                    <IconButton fn={() => setActivePopup(PopupType.HELP)}>
                         <InformationCircleIcon />
                     </IconButton>
                 </Tooltip>
-                <Tooltip content="Settings">
-                    <IconButton fn={() => setSettingsPopupActive(true)}>
+                <Tooltip text="Settings">
+                    <IconButton fn={() => setActivePopup(PopupType.SETTINGS)}>
                         <Cog6ToothIcon />
                     </IconButton>
                 </Tooltip>
@@ -62,7 +60,7 @@ function App() {
                     <Button
                         fn={() => {
                             game.endGame();
-                            setGameOverPopupActive(true)
+                            setActivePopup(PopupType.GAME_OVER);
                         }}
                         classes={clsx(
                             game.gameOver ? 'hidden' : 'visible',
@@ -86,12 +84,12 @@ function App() {
                                 ? 'Game Over'
                                 : `Selection ${game.currAttempt + 1} of 7`
                         }
-                        data={usePlayers(selectedEra).players}
+                        data={players}
                         disabled={game.gameOver}
                         selectPlayer={game.selectPlayer}
                     />
                     <Button
-                        fn={() => setGameOverPopupActive(true)}
+                        fn={() => setActivePopup(PopupType.GAME_OVER)}
                         classes={clsx(
                             game.gameOver ? 'visible' : 'hidden',
                             'justify-self-start place-self-center'
@@ -103,209 +101,13 @@ function App() {
                 <Board board={game.board} correctPlayer={game.correctPlayer} />
             </main>
             <footer>
-                <HelpPopUp />
-                <SettingsPopUp />
-                <ReferencesPopUp />
-                <GameOverPopUp />
+                <HelpPopup activePopup={activePopup} setActivePopup={setActivePopup} />
+                <SettingsPopup game={game} activePopup={activePopup} setActivePopup={setActivePopup} selectedEra={selectedEra} setSelectedEra={setSelectedEra} />
+                <ReferencesPopup activePopup={activePopup} setActivePopup={setActivePopup} />
+                <GameOverPopup game={game} activePopup={activePopup} setActivePopup={setActivePopup} />
             </footer>
         </div>
     )
-
-    function HelpPopUp() {
-        return (
-            <PopUp
-                isVisible={isHelpPopupActive}
-                setIsVisible={setHelpPopupActive}
-            >
-                <div className="font-lg">
-                    <ul>
-                        <li className="p-2">
-                            Your goal is to select the correct player whose name
-                            was on a Bengal's draft card
-                        </li>
-                        <li className="p-2">
-                            You have 7 selections to make the correct pick and
-                            find your draft gem
-                        </li>
-                        <li className="p-2">
-                            A{' '}
-                            <mark className="bg-green-300 dark:bg-green-700">
-                                green box
-                            </mark>{' '}
-                            means that the information is correct
-                        </li>
-                        <li className="p-2">
-                            A{' '}
-                            <mark className="bg-yellow-200 dark:bg-yellow-600">
-                                yellow box
-                            </mark>{' '}
-                            for the year means the correct player is within 5
-                            years
-                        </li>
-                        <li className="p-2">
-                            A{' '}
-                            <mark className="bg-yellow-200 dark:bg-yellow-600">
-                                yellow box
-                            </mark>{' '}
-                            for the position means the correct player is on the
-                            same unit
-                        </li>
-                        <li className="p-2">
-                            A{' '}
-                            <mark className="bg-yellow-200 dark:bg-yellow-600">
-                                yellow box
-                            </mark>{' '}
-                            for the round means the correct player is within 2
-                            rounds
-                        </li>
-                        <li className="p-2">
-                            A{' '}
-                            <mark className="bg-yellow-200 dark:bg-yellow-600">
-                                yellow box
-                            </mark>{' '}
-                            on the picks means the correct player is within 20
-                            picks
-                        </li>
-                    </ul>
-                </div>
-            </PopUp>
-        )
-    }
-
-    function SettingsPopUp() {
-        return (
-            <PopUp
-                isVisible={isSettingsPopupActive}
-                setIsVisible={setSettingsPopupActive}
-            >
-                <div className="grid grid-cols-1 gap-2 place-items-center justify-items-center">
-                    <Button
-                        fn={() => {
-                            game.setNewPlayer()
-                            setSettingsPopupActive(false)
-                        }}
-                    >
-                        new player
-                    </Button>
-                    <Button
-                        fn={() => {
-                            game.resetGame()
-                            setSettingsPopupActive(false)
-                        }}
-                    >
-                        reset board
-                    </Button>
-                    <Dropdown text={selectedEra}>
-                        <DropdownItem fn={() => setSelectedEra(Era.ALL)}>
-                            {Era.ALL}
-                        </DropdownItem>
-                        <DropdownItem
-                            fn={() => setSelectedEra(Era.TWO_THOUSAND_TENS)}
-                        >
-                            {Era.TWO_THOUSAND_TENS}
-                        </DropdownItem>
-                        <DropdownItem
-                            fn={() => setSelectedEra(Era.TWO_THOUSAND)}
-                        >
-                            {Era.TWO_THOUSAND}
-                        </DropdownItem>
-                        <DropdownItem fn={() => setSelectedEra(Era.NINETIES)}>
-                            {Era.NINETIES}
-                        </DropdownItem>
-                        <DropdownItem fn={() => setSelectedEra(Era.EIGHTIES)}>
-                            {Era.EIGHTIES}
-                        </DropdownItem>
-                        <DropdownItem fn={() => setSelectedEra(Era.SEVENTIES)}>
-                            {Era.SEVENTIES}
-                        </DropdownItem>
-                    </Dropdown>
-                </div>
-            </PopUp>
-        )
-    }
-
-    function ReferencesPopUp() {
-        interface LinkProps {
-            href: string
-            children?: ReactNode
-        }
-
-        function Link({ href, children }: LinkProps) {
-            return (
-                <a
-                    className="place-self-start text-orange"
-                    href={href}
-                    target="_blank"
-                    rel="noreferrer"
-                >
-                    {children}
-                </a>
-            )
-        }
-
-        interface LabelProps {
-            children?: ReactNode
-        }
-
-        function Label({ children }: LabelProps) {
-            return <div className="place-self-end">{children}</div>
-        }
-
-        return (
-            <PopUp
-                isVisible={isReferencesPopupActive}
-                setIsVisible={setReferencesPopupActive}
-            >
-                <div className="font-lg grid grid-cols-[1fr_2fr] gap-1 w-fit">
-                    <Label>Inspired By:</Label>
-                    <Link href="https://poeltl.dunk.town/">Poeltl</Link>
-                    <Label>Source Code:</Label>
-                    <Link href="https://github.com/baumace/bengle">
-                        github.com/baumace/bengle
-                    </Link>
-                </div>
-            </PopUp>
-        )
-    }
-
-    function GameOverPopUp() {
-        if (!game.correctPlayer) return null;
-
-        return (
-            <PopUp
-                isVisible={isGameOverPopupActive}
-                setIsVisible={setGameOverPopupActive}
-            >
-                <div className="font-lg">
-                    {game.correctPlayerGuessed ? (
-                        <div>
-                            <p>Draft grade: A+</p>
-                            <p>
-                                You guessed the correct player in {game.currAttempt}{' '}
-                                selections!
-                            </p>
-                        </div>
-                    ) : (
-                        <div>
-                            <p>Draft bust!</p>
-                            <p>
-                                You were not able to select the correct player.
-                            </p>
-                        </div>
-                    )}
-                    <p>
-                        The correct player is {game.correctPlayer.name},{' '}
-                        {game.correctPlayer.position} from {game.correctPlayer.college}
-                    </p>
-                    <p>
-                        Selected with pick #{game.correctPlayer.pick} in round{' '}
-                        {game.correctPlayer.round} of the {game.correctPlayer.year} NFL
-                        Draft
-                    </p>
-                </div>
-            </PopUp>
-        )
-    }
 }
 
 export default App
