+++ ui/src/types/Popup.ts
@@ -1,0 +1,7 @@
+export enum PopupType {
+    GAME_OVER,
+    HELP,
+    SETTINGS,
+    REFERENCES,
+    NONE
+}
